requirements.txt
Flask-RESTful==0.3.9
Flask>=2
gunicorn>=20
PyJWT==2.1.0
